Some changes to HIRS flags
Some changes to the HIRS easyFCDR flags.  suspect_geo and suspect_time
only make sense per line, as they relate to data only happening once per
line.  And most of the flags that were in per_line are really not relevant
for easyFCDR users and will only be used in the full FCDR.